# docs/swagger.yaml
definitions:
...
  dto.AddCompanyIPReq:
    properties:
      ip_addr:
        type: string
    required:
    - ip_addr
    type: object
...
  dto.CloseQuickHustleResp:
    properties:
      data:
        $ref: '#/definitions/dto.QuickHustelBetResData'
      message:
        type: string
    type: object
...
  dto.Company:
    properties:
      created_at:
        type: string
      created_by:
        type: string
      deleted_at:
        type: string
      id:
        type: string
      ip_list:
        items:
          type: string
        type: array
      lockout_duration:
        maximum: 86400
        minimum: 0
        type: integer
      maintenance_mode:
        type: boolean
      maximum_login_attempt:
        maximum: 10
        type: integer
      password_expiry:
        maximum: 365
        minimum: 0
        type: integer
      require_two_factor_authentication:
        type: boolean
      site_name:
        maxLength: 255
        minLength: 1
        type: string
      support_email:
        type: string
      support_phone:
        maxLength: 20
        minLength: 8
        type: string
      updated_at:
        type: string
    required:
    - lockout_duration
    - password_expiry
    - site_name
    - support_email
    - support_phone
    type: object
...
  dto.CreateCompanyReq:
    properties:
      created_by:
        type: string
      ip_list:
        items:
          type: string
        type: array
      lockout_duration:
        maximum: 86400
        minimum: 0
        type: integer
      maintenance_mode:
        type: boolean
      maximum_login_attempt:
        maximum: 10
        minimum: 1
        type: integer
      password_expiry:
        maximum: 365
        minimum: 0
        type: integer
      require_two_factor_authentication:
        type: boolean
      site_name:
        maxLength: 255
        minLength: 1
        type: string
      support_email:
        type: string
      support_phone:
        maxLength: 20
        minLength: 8
        type: string
    required:
    - site_name
    - support_email
    - support_phone
    type: object
  dto.CreateCompanyRes:
    properties:
      created_at:
        type: string
      created_by:
        type: string
      deleted_at:
        type: string
      id:
        type: string
      ip_list:
        items:
          type: string
        type: array
      lockout_duration:
        type: integer
      maintenance_mode:
        type: boolean
      maximum_login_attempt:
        type: integer
      password_expiry:
        type: integer
      require_two_factor_authentication:
        type: boolean
      site_name:
        type: string
      support_email:
        type: string
      support_phone:
        type: string
      updated_at:
        type: string
    type: object
...
  dto.CreateQuickHustelBetRes:
    properties:
      data:
        $ref: '#/definitions/dto.CreateQuickHustelBetResData'
      message:
        type: string
    type: object
  dto.CreateQuickHustelBetResData:
    properties:
      bet_amount:
        type: number
      first_card:
        type: string
      id:
        type: string
      status:
        type: string
      timestamp:
        type: string
      user_id:
        type: string
    type: object
  dto.CreateQuickHustleBetReq:
    properties:
      bet_amount:
        type: number
    type: object
...
  dto.CreateRollDaDiceReq:
    properties:
      bet_amount:
        type: number
      timestamp:
        type: string
      user_guess_end_point:
        type: number
      user_guess_start_point:
        type: number
      won_status:
        type: string
    type: object
  dto.CreateRollDaDiceResp:
    properties:
      data:
        $ref: '#/definitions/dto.RollDaDiceRespData'
      message:
        type: string
    type: object
...
  dto.GetCompaniesRes:
    properties:
      companies:
        items:
          $ref: '#/definitions/dto.Company'
        type: array
      total_pages:
        type: integer
    type: object
...
  dto.GetQuickHustleResp:
    properties:
      data:
        $ref: '#/definitions/dto.QuickHustleGetHistoryData'
      message:
        type: string
    type: object
...
  dto.GetRollDaDiceResp:
    properties:
      data:
        $ref: '#/definitions/dto.GetRollDaDiceRespData'
      message:
        type: string
    type: object
  dto.GetRollDaDiceRespData:
    properties:
      histories:
        items:
          $ref: '#/definitions/dto.RollDaDiceData'
        type: array
      total_pages:
        type: integer
      user_id:
        type: string
    type: object
...
  dto.QuickHustelBetResData:
    properties:
      bet_amount:
        type: number
      first_card:
        type: string
      id:
        type: string
      second_card:
        type: string
      status:
        type: string
      timestamp:
        type: string
      user_guessed:
        type: string
      user_id:
        type: string
      won_amount:
        type: number
      won_status:
        type: string
    type: object
  dto.QuickHustleGetHistoryData:
    properties:
      histories:
        items:
          $ref: '#/definitions/dto.QuickHustelBetResData'
        type: array
      total_pages:
        type: integer
      user_id:
        type: string
    type: object
...
  dto.RollDaDiceData:
    properties:
      bet_amount:
        type: number
      crash_point:
        type: number
      id:
        type: string
      multiplier:
        type: number
      status:
        type: string
      timestamp:
        type: string
      user_guess_end_point:
        type: number
      user_guess_start_point:
        type: number
      user_id:
        type: string
      won_amount:
        type: number
      won_status:
        type: string
    type: object
  dto.RollDaDiceRespData:
    properties:
      bet_amount:
        type: number
      id:
        type: string
      multiplier:
        type: number
      timestamp:
        type: string
      user_guess_end_point:
        type: number
      user_guess_start_point:
        type: number
      user_id:
        type: string
    type: object
...
  dto.SelectQuickHustlePossibilityReq:
    properties:
      id:
        type: string
      user_guess:
        type: string
      user_id:
        type: string
    type: object
...
  dto.UpdateCompanyReq:
    properties:
      id:
        type: string
      lockout_duration:
        maximum: 86400
        minimum: 0
        type: integer
      maintenance_mode:
        type: boolean
      maximum_login_attempt:
        maximum: 10
        minimum: 0
        type: integer
      password_expiry:
        maximum: 365
        minimum: 0
        type: integer
      require_two_factor_authentication:
        type: boolean
      site_name:
        maxLength: 255
        minLength: 1
        type: string
      support_email:
        type: string
      support_phone:
        maxLength: 20
        minLength: 8
        type: string
    type: object
  dto.UpdateCompanyRes:
    properties:
      created_at:
        type: string
      created_by:
        type: string
      deleted_at:
        type: string
      id:
        type: string
      ip_list:
        items:
          type: string
        type: array
      lockout_duration:
        type: integer
      maintenance_mode:
        type: boolean
      maximum_login_attempt:
        type: integer
      password_expiry:
        type: integer
      require_two_factor_authentication:
        type: boolean
      site_name:
        type: string
      support_email:
        type: string
      support_phone:
        type: string
      updated_at:
        type: string
    type: object
...
  response.ErrorResponse:
    properties:
      code:
        description: Code is the error code. It is not status code
        type: integer
      description:
        description: Description is the error description.
        type: string
      field_error:
        description: FieldError is the error detail for each field, if available that
          is.
        items:
          $ref: '#/definitions/response.FieldError'
        type: array
      message:
        description: Message is the error message.
        type: string
      stack_trace:
        description: |-
          StackTrace is the stack trace of the error.
          It is only returned for debugging
        type: string
    type: object
  response.FieldError:
    properties:
      description:
        description: Description is the error description for this field.
        type: string
      name:
        description: Name is the name of the field that caused the error.
        type: string
    type: object
info:
  contact: {}
paths:
...
  /api/admin/companies:
    get:
      consumes:
      - application/json
      description: GetCompanies retrieves a paginated list of companies
      parameters:
      - description: Page number (required)
        in: query
        name: page
        required: true
        type: string
      - description: Items per page (required)
        in: query
        name: per-page
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.GetCompaniesRes'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: GetCompanies
      tags:
      - Admin
    post:
      consumes:
      - application/json
      description: CreateCompany creates a new company
      parameters:
      - description: Create Company Request
        in: body
        name: companyReq
        required: true
        schema:
          $ref: '#/definitions/dto.CreateCompanyReq'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.CreateCompanyRes'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: CreateCompany
      tags:
      - Admin
  /api/admin/companies/{id}:
    delete:
      consumes:
      - application/json
      description: DeleteCompany soft deletes a company by ID
      parameters:
      - description: Company ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: DeleteCompany
      tags:
      - Admin
    get:
      consumes:
      - application/json
      description: GetCompanyByID retrieves a company by its ID
      parameters:
      - description: Company ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Company'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: GetCompanyByID
      tags:
      - Admin
    patch:
      consumes:
      - application/json
      description: UpdateCompany updates an existing company
      parameters:
      - description: Update Company Request
        in: body
        name: companyReq
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateCompanyReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UpdateCompanyRes'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: UpdateCompany
      tags:
      - Admin
  /api/admin/companies/{id}/add-ip:
    patch:
      consumes:
      - application/json
      description: AddIP adds an IP address to a company's allowed list
      parameters:
      - description: Add IP Request
        in: body
        name: addIPReq
        required: true
        schema:
          $ref: '#/definitions/dto.AddCompanyIPReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UpdateCompanyRes'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: AddIP
      tags:
      - Admin
...
  /api/quickhustle/bets:
    get:
      description: GetQuickHustleBetHistory allow user  to get quick hustle bet history.
      parameters:
      - description: 'Bearer <token> '
        in: header
        name: Authorization
        required: true
        type: string
      - description: page type (required)
        in: query
        name: page
        required: true
        type: string
      - description: per_page type (required)
        in: query
        name: per_page
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.GetQuickHustleResp'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: GetQuickHustleBetHistory
      tags:
      - Bet
  /api/quickhustles:
    post:
      consumes:
      - application/json
      description: PlaceQuickHustleBet allow player  to place quick hustle bet
      parameters:
      - description: 'Bearer <token> '
        in: header
        name: Authorization
        required: true
        type: string
      - description: place quick hustle bet  Request
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/dto.CreateQuickHustleBetReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.CreateQuickHustelBetRes'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: PlaceQuickHustleBet
      tags:
      - Bet
  /api/quickhustles/selects:
    post:
      consumes:
      - application/json
      description: UserSelectCard allow player  to send user guess for quick hustle.
      parameters:
      - description: 'Bearer <token> '
        in: header
        name: Authorization
        required: true
        type: string
      - description: guess for quick hustle. req
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/dto.SelectQuickHustlePossibilityReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.CloseQuickHustleResp'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: UserSelectCard
      tags:
      - Bet
  /api/rolldadices:
    get:
      description: GetRollDaDiceHistory allow user  to get roll da dice bets
      parameters:
      - description: 'Bearer <token> '
        in: header
        name: Authorization
        required: true
        type: string
      - description: page type (required)
        in: query
        name: page
        required: true
        type: string
      - description: per_page type (required)
        in: query
        name: per_page
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.GetRollDaDiceResp'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: GetRollDaDiceHistory
      tags:
      - Bet
    post:
      consumes:
      - application/json
      description: CreateRollDaDice allow player  to place roll da dice bet
      parameters:
      - description: 'Bearer <token> '
        in: header
        name: Authorization
        required: true
        type: string
      - description: place roll da dice bet  Request
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/dto.CreateRollDaDiceReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.CreateRollDaDiceResp'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: CreateRollDaDice
      tags:
      - Bet
...
swagger: "2.0"